# HTTP: Reuse `buffer.Bytes()` when decompressing