--- a/pkg/util/http_loki_test.go
+++ b/pkg/util/http_loki_test.go
@@ -1,0 +1,18 @@
+package util
+
+// This file is separated from `http_test.go` to differentiate from the original `http_test.go` file forked from Cortex.
+
+import (
+	"bytes"
+	"testing"
+)
+
+func BenchmarkDecompressFromBuffer(b *testing.B) {
+	buf := bytes.Buffer{}
+	buf.Grow(100)
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		decompressFromBuffer(&buf, 1000, RawSnappy, nil)
+	}
+}