--- a/pkg/util/http_loki_test.go
+++ b/pkg/util/http_loki_test.go
@@ -13,6 +13,6 @@ func BenchmarkDecompressFromBuffer(b *testing.B) {
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		decompressFromBuffer(&buf, 1000, RawSnappy, nil)
+		decompressFromBuffer(&buf, 1000, RawSnappy, nil) //nolint:errcheck
 	}
 }