**What this PR does / why we need it**:
- Modifies `decompressFromBuffer` to call `buffer.Bytes()` a single time.
- Adds a new Benchmark under `http_loki_test.go` file. I had to create this in a separated file to differentiate from the original `http_test.go` forked from Cortex and because `http_test.go` uses package `http_test` which doesn't allow us to test private functions.

Before:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -tags integration,requires_docker -bench ^BenchmarkDecompressFromBuffer$ github.com/grafana/loki/pkg/util

goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/util
cpu: AMD Ryzen 9 7900X 12-Core Processor            
BenchmarkBla-24    	384391501	         3.185 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/grafana/loki/pkg/util	1.547s
```

After:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -tags integration,requires_docker -bench ^BenchmarkDecompressFromBuffer$ github.com/grafana/loki/pkg/util

goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/util
cpu: AMD Ryzen 9 7900X 12-Core Processor            
BenchmarkBla-24    	422789878	         2.716 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/grafana/loki/pkg/util	1.443s
```