@DylanGuedes 2023-04-24
Profiling:
![image](https://user-images.githubusercontent.com/7079397/234078097-11856410-41e2-4d36-9707-a32a3ec0c206.png)